Handles error of user entering non-integer input.
If parse successfully converts the string into a u32 integer,
it will send the number via Ok into the 'match' expression.
Because the Ok value matches the pattern in the first arm,
match will return the num value, which then gets passed to guess.

If the parse does not successfully convert the string into a number, it
returns an Err value that matches the pattern in the second arm of the
match expression. The underscore will match all Err values. The continue
will restart the loop.